% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_categorical_field_labels.R
\name{get_categorical_field_labels}
\alias{get_categorical_field_labels}
\title{Generates a row of code-label pair data for each code in each
categorical variable in a REDCap metatdata file.}
\usage{
get_categorical_field_labels(metadata)
}
\arguments{
\item{metadata}{A REDCap metadata file with columns in the standardized format defined in
https://github.com/SpectrumHealthResearch/REDCapRITS/blob/master/R/R/REDCap_split.r}
}
\value{
A dataframe with the columns, "field_name", "field_type", form_name", "code", and "label".
}
\description{
Generates a row of code-label pair data for each code in each
categorical variable in a REDCap metatdata file.
}
\examples{
get_categorical_field_labels(metadata)
}
