% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_table.R
\name{print_table}
\alias{print_table}
\title{Makes a pretty table with the given data and labels.}
\usage{
print_table(data, main_column_name)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{main_column_name}{A string containing the main column name.}
}
\value{
A kable.
}
\description{
Makes a pretty table with the given data and labels.
}
\examples{
data <- c('Female', 'Female', 'Female', 'Male')
main_column_name <- c('Gender')
print_table(data, main_column_name)
}
