% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_metadata.R
\name{standarize_metadata}
\alias{standarize_metadata}
\title{Renames columns in a REDCap metadata file according to the format defined in
https://github.com/SpectrumHealthResearch/REDCapRITS/blob/master/R/R/REDCap_split.r
and fills out the categorical data for fields of type 'yesno'.}
\usage{
standarize_metadata(metadata)
}
\arguments{
\item{metadata}{A REDCap metadata in REDCap's standard format.}
}
\value{
A REDCap metadata in the SpectrumHealthResearch format
        with filled out categorical data for 'yesno' fields.
}
\description{
Renames columns in a REDCap metadata file according to the format defined in
https://github.com/SpectrumHealthResearch/REDCapRITS/blob/master/R/R/REDCap_split.r
and fills out the categorical data for fields of type 'yesno'.
}
\examples{
standarize_metadata(metadata)
}
