% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_repeating_instruments.R
\name{trim_repeating_instruments}
\alias{trim_repeating_instruments}
\title{Removes repeating instruments.}
\usage{
trim_repeating_instruments(records)
}
\arguments{
\item{records}{A data frame containing the records.}
}
\value{
A data frame containing the records with no repeating elements.
}
\description{
Removes repeating instruments.
}
\examples{
trim_repeating_instruments(Medication_records)
}
