% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_table.R
\name{print_table}
\alias{print_table}
\title{Makes a pretty table with the given data and labels.}
\usage{
print_table(data, columns)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{columns}{A vector containing the column names.}
}
\value{
A kable.
}
\description{
Makes a pretty table with the given data and labels.
}
\examples{
data <- c('Female', 'Female', 'Female', 'Male')
columns <- c('Gender', 'Percentage')
print_table(data, columns)
}
